+++ src/NEventStore/Persistence/PersistStreamsExtensions.cs
@@ -2,6 +2,7 @@ namespace NEventStore.Persistence
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
 
     public static class PersistStreamsExtensions
     {
@@ -54,6 +55,19 @@ namespace NEventStore.Persistence
                 throw new ArgumentNullException(nameof(persistStreams));
             }
             persistStreams.DeleteStream(Bucket.Default, streamId);
+        }
+
+        /// <summary>
+        /// Returns a single commit from any bucket.
+        /// Wrapper for the <see cref="IPersistStreams.GetFromTo(long, long)"/> function in order to
+        /// return a single commit.
+        /// </summary>
+        /// <param name="persistStreams">The IPersistStreams instance.</param>
+        /// <param name="checkpointToken">The checkpoint token that mark the commit to read.</param>
+        /// <returns>A single commit.</returns>
+        public static ICommit GetCommit(this IPersistStreams persistStreams, Int64 checkpointToken)
+        {
+            return persistStreams.GetFromTo(checkpointToken - 1, checkpointToken).SingleOrDefault();
         }
     }
 }