+++ Changelog.md
@@ -4,7 +4,8 @@
 
 - The IPersistStreams interface got some major changes:
 	- Added new GetFromTo(Int64, Int64) and GetFromTo(Strimg, Int64, Int64) methods to the IPersistStreams interface.
-	- Extension methods GetFrom(DateTime) and GetFromTo(DateTime, DateTime) were marked obsolete and will be removed.
+	- Extension methods PersistStreamsExtensions.GetFrom(DateTime) and PersistStreamsExtensions.GetFromTo(DateTime, DateTime) were marked obsolete and will be removed.
+	- A new PersistStreamsExtensions.GetCommit(Int64) method was added to retrieve a single commit [#445](https://github.com/NEventStore/NEventStore/issues/445).
 - PollingClient was moved to its own NEventStore.PollingClient NuGet package [#467](https://github.com/NEventStore/NEventStore/issues/467).
 
 ### Breaking Changes